+++ tests/src/tests/mixnet.rs
@@ -57,7 +57,7 @@ async fn test_one_message(msg_size: usize) {
 #[ignore]
 #[tokio::test(flavor = "multi_thread")]
 async fn mixnet_ten_messages() {
-    let (topology, mut destination_stream) = run_nodes_and_destination_client().await;
+    let (topology, mut destination_stream) = tests::run_nodes_and_destination_client().await;
 
     let mut msg = [0u8; 100 * 1024];
     rand::thread_rng().fill_bytes(&mut msg);
@@ -90,93 +90,4 @@ async fn mixnet_ten_messages() {
 
     let elapsed = Instant::now().checked_duration_since(start_time).unwrap();
     println!("ELAPSED: {elapsed:?}");
-}
-
-async fn run_nodes_and_destination_client() -> (
-    MixnetTopology,
-    impl Stream<Item = Result<Vec<u8>, MixnetClientError>> + Send,
-) {
-    let config1 = MixnetNodeConfig {
-        listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 7777)),
-        client_listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 7778)),
-        ..Default::default()
-    };
-    let config2 = MixnetNodeConfig {
-        listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8777)),
-        client_listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8778)),
-        ..Default::default()
-    };
-    let config3 = MixnetNodeConfig {
-        listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 9777)),
-        client_listen_address: SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 9778)),
-        ..Default::default()
-    };
-
-    let mixnode1 = MixnetNode::new(config1.clone());
-    let mixnode2 = MixnetNode::new(config2.clone());
-    let mixnode3 = MixnetNode::new(config3.clone());
-
-    let topology = MixnetTopology {
-        layers: vec![
-            Layer {
-                nodes: HashMap::from([(
-                    mixnode1.id(),
-                    Node {
-                        address: config1.listen_address,
-                        public_key: mixnode1.public_key(),
-                    },
-                )]),
-            },
-            Layer {
-                nodes: HashMap::from([(
-                    mixnode2.id(),
-                    Node {
-                        address: config2.listen_address,
-                        public_key: mixnode2.public_key(),
-                    },
-                )]),
-            },
-            Layer {
-                nodes: HashMap::from([(
-                    mixnode3.id(),
-                    Node {
-                        address: config3.listen_address,
-                        public_key: mixnode3.public_key(),
-                    },
-                )]),
-            },
-        ],
-    };
-
-    // Run all MixnetNodes
-    tokio::spawn(async move {
-        let res = mixnode1.run().await;
-        assert!(res.is_ok());
-    });
-    tokio::spawn(async move {
-        let res = mixnode2.run().await;
-        assert!(res.is_ok());
-    });
-    tokio::spawn(async move {
-        let res = mixnode3.run().await;
-        assert!(res.is_ok());
-    });
-
-    // Wait until mixnodes are ready
-    // TODO: use a more sophisticated way
-    tokio::time::sleep(Duration::from_secs(1)).await;
-
-    // Run a MixnetClient only for the MixnetNode in the exit layer.
-    // According to the current implementation,
-    // one of mixnodes the exit layer always will be selected as a destination.
-    let client = MixnetClient::new(
-        MixnetClientConfig {
-            mode: MixnetClientMode::SenderReceiver(config3.client_listen_address),
-            topology: topology.clone(),
-        },
-        OsRng,
-    );
-    let client_stream = client.run().await.unwrap();
-
-    (topology, client_stream)
 }
